started work on getting a unit to act over the network

+++ Assets/Resorces/Scripts/GridBlock.cs
@@ -315,7 +315,7 @@ public class GridBlock : MonoBehaviour,IPointerDownHandler{
 	}
 
 	public void spawUnitPlayerFromNetwork(string unitName){//change Alliance to team.getTeamIndex
-		Player.Instance.thisPlayersNetworkHelper.Cmd_SendUnitSpawnEventToServer(unitName, gridlocation.x, gridlocation.y, (byte)Player.Instance.playerAlliance);
+		Player.Instance.thisPlayersNetworkHelper.Cmd_SendUnitSpawnEventToServer(unitName, (ushort)gridlocation.x, (ushort)gridlocation.y, (byte)Player.Instance.playerAlliance);
 	}
 	#endregion
